MDEV-31949 II.1 slow parallel replication of user xa
This is the 1st commit of the II part of the 3 parts branch.
The part I refactors the binary logging of XA to integrate
it properly with the Binlog-Group-Commit module.
The part III is concerned with an optimal durable write
policy by the user XA-PREPARE,COMMIT,ROLLBACK.

Introduced by MDEV-742 XA-Prepare group of events

  XA START xid
  ...
  XA END xid
  XA PREPARE xid

and its XA-"complete" terminator

  XA COMMIT or
  XA ROLLBACK

that constitute a binlog image of a user XA transaction
have now been made distributed Round-Robin across slave parallel workers.
The original hash-based policy was reasonably good in identifying
the XA-PREPARE worker by a respective XA-"COMPLETE".
Yet the hash method was proven to attribute to execution latency
through creating a big - many times larger than the size of the worker
pool - queue of binlog-ordered transactions to commit.

Acronyms used in the commit message and patch comments:

  XAP := XA-Prepare event or the whole prepared XA group of events
  XAC := XA-"complete", which is a solitary group of events,
         sometimes specifically XA-Commit.

NOTABLE CHANGES in this commit are done mostly to
the parallel slave module.

The parallel slave driver thread now maintains a list XAP:s currently
in scheduling. Its purpose is to avoid parallel execution of XA:s
with duplicate XIDs (unlikely, but that's the user's right).
The list is arranged as a sliding window with the size of the slave
worker pool which is sufficient to resolve any XAP_i -> XAP_i+k
dependency. E.g having  4 worker threads XAP_5(xid) can be safely
distributed when its XID did not occur in XAP:s indexed 2,3,4; it
could in XAP_1, but this would be safe because both would belong
to the same worker.

The current commit also tracks XAP_i -> XAC_i+k dependency between
the prepared and the commit parts of the same transaction and does
so rather conservatively. This is optimized away in the following commit.
The dependency XAP -xid-> XA[PC] tracker works as follows.
When a new XAP (or XAC in this commit) enters the replication applier its XID
is compared against ones in the window. If the current one duplicates,
e.g when it's XAC(xid) that comes in after XAP(xid), then
XAC would be 'waiting for prior transaction to commit'.
In this example among the prior transactions one would necessarily be
the XAC's "parent" XA-PREPARE part of the whole XA transaction.

OTHER CHANGES,
to
  ha_close_connection()
are done to keep the binlog hton connection alive for the end of
XA-prepare's binlogging on the parallel slave.
Other than the binlog hton:s are disconnected from the slave
transaction handler in TC_LOG::run_commit_ordered() which makes
the prepared transaction engine objects available to
binlog-commit-order followers.

UNRELATED
to
  rpl_xa_prepare_gtid_fail.test
added expected warning suppression that MDEV-31038 missed out.

Thanks to Brandon Nesterenko at mariadb.com for initial review and
a lot of creative efforts to advance with this work!
Enormous kudos to Kristian Nielsen for thorough insightful reviews!
Roel Van de Paar helped to improved from the testing.