MDEV-31949 III.2 last XA in binlog group invokes engine durable write
This commit constricts the previous commit's no-flush policy for XAP
and XAC binlog-group-members. The last of them in a binlog group
invokes the engine durable write, while ones in the middle of the
group don't. This measure ensures the pre-existing crash-recovery
behavior has not changed by the fixes branch.
Specifically a crash in the middle of the binlog group such that
an XA operation got written in binlog (sync_binlog = 1 is assumed)
but its engine branch was not updated durably is covered
by MDEV-21469. A crash crash after the last XA operation is engine
is durable is covered and tested here.

A new
  binlog.binlog_xa_mdev-31949
and existing
  binlog.binlog_xa_prepared_disconnect and
  binlog.binlog_xa_prepared_disconnect
demonstrate that when an XA command in {PREPARE,COMMIT,ROLLBACK} succeeds

       -> XA command xid
       <= OK

it's written persistently prior to OK so after crash and restart the
respective command's effects are proven to last.

E.g for the PREPARE the following

       -> XA RECOVER

must return the xid

       <= xid

Implementation-wise the last xa prepare,commit or rollback of a
binlogging transaction group is marked with its binlog offset zero by
the binlog group commit leader.  In the Engine the zero offset is now
detected to skip the durable write by such transaction.